+++ arclasp/__init__.py
@@ -37,7 +37,7 @@ from arclasp.exceptions import (
     ArclaspVerificationError,
 )
 
-__version__ = "0.1.0b2"
+__version__ = "0.1.0b3"
 
 __all__ = [
     "init",